feature: create BlockRenderController and move render logic
